Whitespace changes

+++ TexturePlotter.cpp
@@ -199,7 +199,7 @@ void TexturePlotter::UpdateTable() {
 		case 0: {// Cell area
 			for (size_t i = 0; i < w; i++) {
 				for (size_t j = 0; j < h; j++) {
-					double val = worker->GetGeometry()->GetPhysicalValue(selFacet, PhysicalMode::CellArea, 1.0, 1.0, 1.0, i + j * w, NULL).value;
+					double val = worker->GetGeometry()->GetPhysicalValue(selFacet, PhysicalMode::CellArea, 1.0, 1.0, 1.0, (int)(i + j*w), NULL).value;
 					sprintf(tmp, "%g", val);
 					if (val > maxValue) {
 						maxValue = val;
@@ -222,7 +222,7 @@ void TexturePlotter::UpdateTable() {
 						for (size_t j = 0; j < h; j++) {
 							//int tSize = selFacet->wp.texWidth*selFacet->wp.texHeight;
 
-							double val = worker->GetGeometry()->GetPhysicalValue(selFacet, PhysicalMode::MCHits, 1.0, 1.0, 1.0, i + j * w, (BYTE*)texture).value;
+							double val = worker->GetGeometry()->GetPhysicalValue(selFacet, PhysicalMode::MCHits, 1.0, 1.0, 1.0, (int)(i + j*w), (BYTE*)texture).value;
 							if (val > maxValue) {
 								maxValue = (double)val;
 								maxX = i; maxY = j;
@@ -256,7 +256,7 @@ void TexturePlotter::UpdateTable() {
 					double moleculesPerTP = (worker->wp.sMode == MC_MODE) ? mApp->worker.GetMoleculesPerTP(worker->displayedMoment) : 1.0;
 					for (size_t i = 0; i < w; i++) {
 						for (size_t j = 0; j < h; j++) {
-							double val = worker->GetGeometry()->GetPhysicalValue(selFacet, PhysicalMode::ImpingementRate, moleculesPerTP,1.0, worker->wp.gasMass, i+j*w, (BYTE*)texture).value;
+							double val = worker->GetGeometry()->GetPhysicalValue(selFacet, PhysicalMode::ImpingementRate, moleculesPerTP,1.0, worker->wp.gasMass, (int)(i + j * w), (BYTE*)texture).value;
 							if (val > maxValue) {
 								maxValue = val;
 								maxX = i; maxY = j;
@@ -288,7 +288,7 @@ void TexturePlotter::UpdateTable() {
 
 					for (size_t i = 0; i < w; i++) {
 						for (size_t j = 0; j < h; j++) {
-							double rho = worker->GetGeometry()->GetPhysicalValue(selFacet, PhysicalMode::ParticleDensity, moleculesPerTP, densityCorrection, worker->wp.gasMass, i + j * w, (BYTE*)texture).value;
+							double rho = worker->GetGeometry()->GetPhysicalValue(selFacet, PhysicalMode::ParticleDensity, moleculesPerTP, densityCorrection, worker->wp.gasMass, (int)(i + j*w), (BYTE*)texture).value;
 							if (rho > maxValue) {
 								maxValue = rho;
 								maxX = i; maxY = j;
@@ -325,7 +325,7 @@ void TexturePlotter::UpdateTable() {
 							/*double v_avg = 2.0*(double)texture[i + j*w].count / texture[i + j*w].sum_1_per_ort_velocity;
 							double imp_rate = texture[i + j*w].count / (selFacet->mesh[i + j*w].area*(selFacet->wp.is2sided ? 2.0 : 1.0))*dCoef;
 							double rho = 4.0*imp_rate / v_avg;*/
-							double rho_mass = worker->GetGeometry()->GetPhysicalValue(selFacet, PhysicalMode::GasDensity, moleculesPerTP, densityCorrection, worker->wp.gasMass, i + j * w, (BYTE*)texture).value;
+							double rho_mass = worker->GetGeometry()->GetPhysicalValue(selFacet, PhysicalMode::GasDensity, moleculesPerTP, densityCorrection, worker->wp.gasMass, (int)(i + j*w), (BYTE*)texture).value;
 							if (rho_mass > maxValue) {
 								maxValue = rho_mass;
 								maxX = i; maxY = j;
@@ -357,7 +357,7 @@ void TexturePlotter::UpdateTable() {
 					for (size_t i = 0; i < w; i++) {
 						for (size_t j = 0; j < h; j++) {
 
-							double p = worker->GetGeometry()->GetPhysicalValue(selFacet, PhysicalMode::Pressure, moleculesPerTP, 1.0, worker->wp.gasMass, i + j * w, (BYTE*)texture).value;
+							double p = worker->GetGeometry()->GetPhysicalValue(selFacet, PhysicalMode::Pressure, moleculesPerTP, 1.0, worker->wp.gasMass, (int)(i + j*w), (BYTE*)texture).value;
 							if (p > maxValue) {
 								maxValue = p;
 								maxX = i; maxY = j;
@@ -387,7 +387,7 @@ void TexturePlotter::UpdateTable() {
 					TextureCell *texture = (TextureCell *)((BYTE *)buffer + (selFacet->sh.hitOffset + facetHitsSize + profSize + mApp->worker.displayedMoment*w*h * sizeof(TextureCell)));
 					for (size_t i = 0; i < w; i++) {
 						for (size_t j = 0; j < h; j++) {
-							double val = worker->GetGeometry()->GetPhysicalValue(selFacet, PhysicalMode::AvgGasVelocity, 1.0, 1.0, 1.0, i + j * w, (BYTE*)texture).value;
+							double val = worker->GetGeometry()->GetPhysicalValue(selFacet, PhysicalMode::AvgGasVelocity, 1.0, 1.0, 1.0, (int)(i + j*w), (BYTE*)texture).value;
 							if (val > maxValue) {
 								maxValue = val;
 								maxX = i; maxY = j;
@@ -419,7 +419,7 @@ void TexturePlotter::UpdateTable() {
 					for (size_t i = 0; i < w; i++) {
 						for (size_t j = 0; j < h; j++) {
 							if (selFacet->sh.countDirection) {
-								Vector3d v_vect = worker->GetGeometry()->GetPhysicalValue(selFacet, PhysicalMode::GasVelocityVector, 1.0, 1.0, 1.0, i + j * w, (BYTE*)dirs).vect;
+								Vector3d v_vect = worker->GetGeometry()->GetPhysicalValue(selFacet, PhysicalMode::GasVelocityVector, 1.0, 1.0, 1.0, (int)(i + j*w), (BYTE*)dirs).vect;
 								sprintf(tmp, "%g,%g,%g",
 									v_vect.x, v_vect.y, v_vect.z);
 								double length = v_vect.Norme();
@@ -457,7 +457,7 @@ void TexturePlotter::UpdateTable() {
 					for (size_t i = 0; i < w; i++) {
 						for (size_t j = 0; j < h; j++) {
 							if (selFacet->sh.countDirection) {
-								size_t count = worker->GetGeometry()->GetPhysicalValue(selFacet, PhysicalMode::NbVelocityVectors, 1.0, 1.0, 1.0, i + j * w, (BYTE*)dirs).count;
+								size_t count = worker->GetGeometry()->GetPhysicalValue(selFacet, PhysicalMode::NbVelocityVectors, 1.0, 1.0, 1.0, (int)(i + j*w), (BYTE*)dirs).count;
 								sprintf(tmp, "%zd", count);		
 								double countEq = (double)count;
 								if (countEq > maxValue) {
